[XLA] Partition computations by directly traversing the call-graph.
Currently, the SPMD partitioner traverses the call-graph (for kControlFlow call sites) by traversing the instructions of each computation and partitioning called computations on the fly. This refactors it to explicitly walk the call-graph instead. This is not expected to have any effect when the call-graph is a tree, and is a preparatory CL for allowing general DAGs.

PiperOrigin-RevId: 797278088